+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### 0.6.5 (develop)
 
 **- Make Rocketeer prompt for both server and SCM credentials if they're not stored**
+**- `artisan deploy` now deploys the project if the `--version` flat is not passed**
 - Make Rocketeer forget invalid credentials provided by prompt
 - Fix a bug where incorrect SCM urls would be generated